fix %p

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: asouta <asouta@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 16:48:11 by asouta            #+#    #+#             */
-/*   Updated: 2021/11/03 16:48:15 by asouta           ###   ########.fr       */
+/*   Updated: 2022/04/15 17:05:37 by asouta           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ char	**ft_split(char const *s, char c);
 char	*ft_itoa(int n);
 char	*ft_utoa(unsigned int n);
 char	*ft_ultoa(unsigned long n);
-char	*ft_ultoxtoa(unsigned int u, char *base);
+char	*ft_ultoxtoa(unsigned long long u, char *base);
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 void	ft_striteri(char *s, void (*f)(unsigned int, char *));
 void	ft_putchar_fd(char c, int fd);
